Add the basis of scrim layout for create rule
For this layout, we need to be able to control the scrim color in the
statusBar area, as well as the toolbar background color.

This is the first part of this feature. It deals with the pain of
WindowInsets in order to arrange this layout just like we need it.
 We still need to listen torecyclerView scroll events and animate the
color according to its scroll value but that should be way easier.

Also, I have started testing groupie since it seems a good fit for this
project : quick easy adapter writing based on databinding