Gamma Evaluator
ArgEvaluator is not that great for color animations. It creates very
ugly color jumps that just look wrong.

This is because RGB colors are gamma encoded, their variation follow a
power law, so a basic linear interpolation does not work. It leads to
ugly color jumps.

We can solve this by converting the color components to linear encoded
values, interpolate and then bring them back.